+++ src/tokens/components/legal-footer.ts
@@ -1,0 +1,4 @@
+import legalFooter from '../../../tokens/component/legal-footer.json';
+import { flattenTokens } from '../utils';
+
+export const legalFooterTokens = flattenTokens(legalFooter as never);